+++ src/main/java/dte/tzevaadomnotifier/tzevaadomlisteners/SoundNotifierListener.java
@@ -1,4 +1,4 @@
-package dte.tzevaadomnotifier.notifiers;
+package dte.tzevaadomnotifier.tzevaadomlisteners;
 
 import org.bukkit.Bukkit;
